Move validation methods to a separate validator package
