+++ pkg/validator/doc.go
@@ -1,0 +1,2 @@
+// Package validator provides methods for validation
+package validator
